refactor: 캐시 조회 계층 분리
