+++ src/main/java/com/prography/lighton/member/application/CompleteMemberProfileService.java
@@ -11,7 +11,7 @@ import com.prography.lighton.member.domain.repository.TemporaryMemberRepository;
 import com.prography.lighton.member.exception.DuplicateMemberException;
 import com.prography.lighton.member.presentation.dto.request.CompleteMemberProfileRequestDTO;
 import com.prography.lighton.member.presentation.dto.response.CompleteMemberProfileResponseDTO;
-import com.prography.lighton.region.domain.resolver.RegionResolver;
+import com.prography.lighton.region.infrastructure.cache.RegionCache;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -24,14 +24,14 @@ public class CompleteMemberProfileService implements CompleteMemberProfileUseCas
     private final TemporaryMemberRepository temporaryMemberRepository;
     private final MemberRepository memberRepository;
 
-    private final RegionResolver regionResolver;
+    private final RegionCache regionCache;
     private final TokenProvider tokenProvider;
 
     @Override
     public CompleteMemberProfileResponseDTO completeMemberProfile(final Long temporaryMemberId,
                                                                   final CompleteMemberProfileRequestDTO request) {
         TemporaryMember temporaryMember = getTemporaryMember(temporaryMemberId);
-        RegionInfo preferredRegion = regionResolver.resolve(request.regionCode());
+        RegionInfo preferredRegion = regionCache.resolve(request.regionCode());
         Phone phone = validatePhoneDuplicate(request.phone());
         MarketingAgreement marketingAgreement = toMarketingAgreement(request);
